+++ service/src/main/kotlin/io/provenance/explorer/domain/extensions/CoinExtensions.kt
@@ -11,13 +11,20 @@ fun BigDecimal.stringfy() = this.stripTrailingZeros().toPlainString()
 
 fun BigDecimal.toCoinStr(denom: String) = CoinStr(this.stringfy(), denom)
 
-fun String.toDecCoin() = BigDecimal(this.toBigInteger(), 18).stripTrailingZeros().toPlainString()
-
-fun String.toPercentage() =
-    BigDecimal(this.toBigInteger(), 18).multiply(BigDecimal(100)).stripTrailingZeros().toPlainString() + "%"
+fun String.toDecCoin() = this.toDecimal().toPlainString()
 
 fun List<CoinOuterClass.Coin>.toProtoCoin() =
     this.firstOrNull() ?: CoinOuterClass.Coin.newBuilder().setAmount("0").setDenom("").build()
 
 fun ServiceOuterClass.GetTxResponse.toCoinStr() =
     this.tx.authInfo.fee.amountList.toProtoCoin().let { coin -> CoinStr(coin.amount, coin.denom) }
+
+// Math extensions
+fun String.toPercentage() =
+    BigDecimal(this.toBigInteger(), 18).multiply(BigDecimal(100)).stripTrailingZeros().toPlainString() + "%"
+
+fun String.toDecimal() = BigDecimal(this.toBigInteger(), 18).stripTrailingZeros()
+
+fun Double.toPercentage() = "${this * 100}%"
+
+fun List<Int>.avg() = this.sum() / this.size